requirements.txt
oracledb
python-pptx
langchain_core
langchain_openai
langchain_community
matplotlib
chromadb
vanna
requests
gradio
vanna[mssql]
flask
python-dotenv
openpyxl
streamlit
plotly
numpy
pandas
matplotlib
python-pptx
dashscope
